Add helpers for existance of aggragations
Calling any of the bucket or metric aggs functions on HasAggregations
with an invalid name, will result in a no-such-key exception. The
assumption seems to be that with an HasAggregation at hand, it is
always known which aggragations it will contain. This seems a bit
too strong an assumption, as such these helpers to figure out which
aggregations names are available.